requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.5.5.64
# opencv-python==4.4.0.46
Pillow==9.3.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
six==1.15.0
imagehash==4.2.1
scikit-image==0.19.3
scikit-learn~=0.24.2
scipy~=1.7.1
pandas==1.3.5
# tensorflow==
sklearn==
# sklearn
# opencv-contrib-python==4.6.0.66